feat!: constraint handling is moved to Iry methods
The original API that presumed that `user.save` would just populate the errors ends up in a lot of corner cases:
- Should the `after_` callbacks still executes?
- Should it be possible to keep running SQL queries after a constraint is fetched? What if these are writing operations?
- `create!` and `update!` wouldn't raise anymore
- monkey patching `create_or_update` would be required
- How to handle `find_or_create_by`?

Overall, this seemed to break the API. When a constraint is raised, transactions must be stopped, as such any code afterward must be halted.

Decision is made to transition to an API that explicitly states when constraints are handled, using `handle_constraints`.
Two helper methods are provided: `.save(model)` and `.save!(model)` to ease usage of the more verbose `handle_constraints`.